After implementation of the influence of advection of theta-fluctuations on buoyancy in  fullD_explicit_update() I see the first consistent results when implicit gravity is switched on everywhere. Still running at rather small time steps, though.

+++ RKLM_Reference/Management/memory.c
@@ -33,7 +33,7 @@ void update(
 	
     const double grav = ud.gravity_strength[1];
     const double Msq  = ud.Msq;
-    const double impg = ud.implicit_gravity_theta;
+    const int    impg = ud.implicit_gravity_theta;
     
     double drho, drhoe_total;
   	int ndim = elem->ndim;
